Buy NFT test

+++ contracts/NftMarket.sol
@@ -1,41 +1,36 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.4.22 <0.9.0;
 
-import "@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
-import "@openzeppelin/contracts/utils/Counters.sol";
+import '@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol';
+import '@openzeppelin/contracts/utils/Counters.sol';
 
 contract NftMarket is ERC721URIStorage {
   using Counters for Counters.Counter;
 
   struct NftItem {
-    uint tokenId;
-    uint price;
+    uint256 tokenId;
+    uint256 price;
     address creator;
     bool isListed;
   }
 
-  uint public listingPrice = 0.025 ether;
+  uint256 public listingPrice = 0.025 ether;
 
   Counters.Counter private _listedItems;
   Counters.Counter private _tokenIds;
 
   mapping(string => bool) private _usedTokenURIs;
-  mapping(uint => NftItem) private _idToNftItem;
+  mapping(uint256 => NftItem) private _idToNftItem;
 
-  event NftItemCreated (
-    uint tokenId,
-    uint price,
-    address creator,
-    bool isListed
-  );
+  event NftItemCreated(uint256 tokenId, uint256 price, address creator, bool isListed);
 
-  constructor() ERC721("CreaturesNFT", "CNFT") {}
+  constructor() ERC721('CreaturesNFT', 'CNFT') {}
 
-  function getNftItem(uint tokenId) public view returns (NftItem memory) {
+  function getNftItem(uint256 tokenId) public view returns (NftItem memory) {
     return _idToNftItem[tokenId];
   }
 
-  function listedItemsCount() public view returns (uint) {
+  function listedItemsCount() public view returns (uint256) {
     return _listedItems.current();
   }
 
@@ -43,14 +38,14 @@ contract NftMarket is ERC721URIStorage {
     return _usedTokenURIs[tokenURI] == true;
   }
 
-  function mintToken(string memory tokenURI, uint price) public payable returns (uint) {
-    require(!tokenURIExists(tokenURI), "Token URI already exists");
-    require(msg.value == listingPrice, "Price must be equal to listing price");
+  function mintToken(string memory tokenURI, uint256 price) public payable returns (uint256) {
+    require(!tokenURIExists(tokenURI), 'Token URI already exists');
+    require(msg.value == listingPrice, 'Price must be equal to listing price');
 
     _tokenIds.increment();
     _listedItems.increment();
 
-    uint newTokenId = _tokenIds.current();
+    uint256 newTokenId = _tokenIds.current();
 
     _safeMint(msg.sender, newTokenId);
     _setTokenURI(newTokenId, tokenURI);
@@ -60,14 +55,12 @@ contract NftMarket is ERC721URIStorage {
     return newTokenId;
   }
 
-  function buyNft(
-    uint tokenId
-  ) public payable {
-    uint price = _idToNftItem[tokenId].price;
+  function buyNft(uint256 tokenId) public payable {
+    uint256 price = _idToNftItem[tokenId].price;
     address owner = ERC721.ownerOf(tokenId);
 
-    require(msg.sender != owner, "You already own this NFT");
-    require(msg.value == price, "Please submit the asking price");
+    require(msg.sender != owner, 'You already own this NFT');
+    require(msg.value == price, 'Please submit the asking price');
 
     _idToNftItem[tokenId].isListed = false;
     _listedItems.decrement();
@@ -76,18 +69,10 @@ contract NftMarket is ERC721URIStorage {
     payable(owner).transfer(msg.value);
   }
 
-  function _createNftItem(
-    uint tokenId,
-    uint price
-  ) private {
-    require(price > 0, "Price must be at least 1 wei");
+  function _createNftItem(uint256 tokenId, uint256 price) private {
+    require(price > 0, 'Price must be at least 1 wei');
 
-    _idToNftItem[tokenId] = NftItem(
-      tokenId,
-      price,
-      msg.sender,
-      true
-    );
+    _idToNftItem[tokenId] = NftItem(tokenId, price, msg.sender, true);
 
     emit NftItemCreated(tokenId, price, msg.sender, true);
   }
